Don't trim pre-context for patches at end of file
Interdiff trims the pre-context to match the size of the post-context.
As far as I can tell, the reason for this is that patch(1) assumes that
fewer post-context lines means that the hunk is at the end of a file.

But when it actually *is* at the end of a file, this causes a problem.
The pre-context is completely removed to match the post-context size,
resulting in the removal of valuable pre-context.

Add a check to prevent this situation by ensuring that the post-context
size is greater than zero before trimming the pre-context.

Also fix up some test cases which were relying on the old behavior.

This fixes https://fedorahosted.org/patchutils/ticket/32.