Make patch arguments more extensible in apply_patch()
It's difficult to conditionally add additional arguments to the patch
execution in apply_patch() because they are placed within a compound
literal array.

Make the arguments more extensible by creating a local array and an index
variable to place the next argument into the array. This way, it's much
easier to change the number of arguments provided at runtime.