Implement a fuzzy diffing option for interdiff
This implements a --fuzzy option to make interdiff perform a fuzzy
comparison between two diffs. This is very helpful, for example, for
comparing a backport patch to its upstream source patch to assist a human
reviewer in verifying the correctness of the backport.

To achieve fuzzy diffing: Instead of creating a different dummy original
file for p1 and p2 and applying p1 to p2's original file and p2 to p1's
original file, use p1's original file for both cases when fuzzy diffing.

This way, interdiff can rely upon the patch command to fuzz the line offset
and context differences between hunks in p1 and p2. And then the final diff
command at the end will show the remaining differences using the same base
file, so it won't report hunks as differing just because of differing line
offsets. The diff command can't do any fuzzy diffing, hence why we need to
rely on the patch command for this capability.

The fuzzy diffing option also handles rejected hunks by showing them in the
output as a differing hunk. And it tries to avoid *additionally* showing
the *reverse* of the rejected hunks by applying the rejected hunks with
maximum fuzz, so that the final diff command doesn't show them.