Fix content skipping for patch2 in index_patch_generic()
When an @@ line isn't immediately after the +++ line in patch2, the next
line is checked from the top of the loop which tries to search for a +++
line again, even though the +++ was already found. This results in the +++
not being found again and thus a spurious error that patch2 is empty.

Fix this by making the patch2 case loop over the next line until either an
@@ is found or the patch is exhausted.