Begin fixing the direction of delta differences in fuzzy mode
Fuzzy mode output is incoherent because context differences show up as
going from patch1 -> patch2, while delta differences show up in the
opposite direction: patch2 -> patch1.

This makes it rather impossible to tell which patch file a +/- line
originates from, not to mention that the diff itself is totally nonsense in
terms of the actual code changed by the patches.

Bring interdiff part of the way there to fixing this issue, by eliminating
delta differences from the compared files. Note that this makes fuzzy mode
only do context diffing, which is fixed in the subsequent commits.